+++ example/lib/sample/sample_app_bar.dart
@@ -7,7 +7,7 @@ class SampleAppBar extends StatefulWidget {
 }
 
 class _SampleAppBarState extends State<SampleAppBar>
-    implements ScaffoldFactoryButtonsBehavior {
+    implements ScaffoldFactoryBehaviors {
   final _scaffoldKey = GlobalKey<ScaffoldState>();
   ScaffoldFactory _scaffoldFactory;
   MaterialPalette _sampleColorPalette = MaterialPalette(
@@ -33,8 +33,11 @@ class _SampleAppBarState extends State<SampleAppBar>
   }
 
   void _initScaffoldFactory() {
-    _scaffoldFactory = ScaffoldFactory(_scaffoldKey, _sampleColorPalette);
-    _scaffoldFactory.buttonsBehavior = this;
+    _scaffoldFactory = ScaffoldFactory(
+      scaffoldKey: _scaffoldKey,
+      materialPalette: _sampleColorPalette,
+    );
+    _scaffoldFactory.scaffoldFactoryBehavior = this;
 
     _scaffoldFactory.init(
       backgroundType: BackgroundType.normal,
@@ -151,5 +154,10 @@ class _SampleAppBarState extends State<SampleAppBar>
   @override
   void onFloatingActionButtonPressed() {
     print("Scaffold Factory => onFloatingActionButtonPressed()");
+  }
+
+  @override
+  Future onEventBusMessageReceived(event) async {
+    print("ScaffoldFactory: Event Received");
   }
 }