+++ lib/event_bus.dart
@@ -1,0 +1,4 @@
+import 'package:event_bus/event_bus.dart';
+
+/// The global [EventBus] object.
+EventBus eventBus = EventBus();